refactoring and style updates
