+++ src/components/icons/SortIcon.tsx
@@ -8,7 +8,10 @@ export default function SortIcon(props) {
             viewBox={props.viewBox}
             width={props.width}>
             <path d="M0 0h24v24H0V0z" fill="none" />
-            <path d="M3 18h6v-2H3v2zM3 6v2h18V6H3zm0 7h12v-2H3v2z" />
+            <path
+                d="M3 18h6v-2H3v2zM3 6v2h18V6H3zm0 7h12v-2H3v2z"
+                fill="currentColor"
+            />
         </svg>
     );
 }
